Update and rename cowandbull to cowandbull.py
Converting entry to integer and to string removes the '0' entry especially if its the first few digits, resolved.
Both the bull and cow portion need a break statement from the for loop, so that the loop will not do duplicate matching. This has previously resulted in extra number of cows.
Userlist shifted into the while loop so that it will reset with every entry from the user.

Code should work this time!